refactored user (#31)
refactored /src/auth/user.schema to /r/auth/UserAuthSchema
