+++ src/activity/services/activity/activity.service.spec.ts
@@ -4,7 +4,7 @@ import mongoose, { Model } from 'mongoose';
 import { getModelToken } from '@nestjs/mongoose';
 import { Activity } from '../../schemas/activity.schema';
 import { CreateActivityDto } from '../../dto/create-activity.dto';
-import { User } from '../../../auth/schemas/user.schema';
+import { User } from '../../../auth/schemas/userAuthSchema';
 import { BadRequestException, NotFoundException } from '@nestjs/common';
 import createMockActivity from '../../../../test/mock-data/createMockActivity';
 import mockActivityFromDB from '../../../../test/mock-data/returned-mock-activity';
