+++ src/activity/services/activity/activity.service.ts
@@ -7,7 +7,7 @@ import { InjectModel } from '@nestjs/mongoose';
 import mongoose, { Model } from 'mongoose';
 import { Activity } from '../../schemas/activity.schema';
 import { Query } from 'express-serve-static-core';
-import { User } from '../../../auth/schemas/user.schema';
+import { User } from '../../../auth/schemas/userAuthSchema';
 
 @Injectable()
 export class ActivityService {
