+++ src/auth/jwt.strategy.ts
@@ -3,7 +3,7 @@ import { InjectModel } from '@nestjs/mongoose';
 import { PassportStrategy } from '@nestjs/passport';
 import { Model } from 'mongoose';
 import { Strategy, ExtractJwt } from 'passport-jwt';
-import { User } from './schemas/user.schema';
+import { User } from './schemas/userAuthSchema';
 
 @Injectable()
 export class JwtStrategy extends PassportStrategy(Strategy) {
