+++ src/auth/schemas/userAuthSchema.ts
@@ -23,4 +23,4 @@ export class User extends Document {
   role: Role;
 }
 
-export const UserSchema = SchemaFactory.createForClass(User);
+export const UserAuthSchema = SchemaFactory.createForClass(User);
