+++ src/user/user.module.ts
@@ -2,10 +2,10 @@
 // import { MongooseModule } from '@nestjs/mongoose';
 // import { UserController } from './controllers/user/user.controller';
 // import { UserService } from './services/user/user.service';
-// import { UserSchema } from './schemas/user.model';
+// import { UserAuthSchema } from './schemas/user.model';
 //
 // @Module({
-//   imports: [MongooseModule.forFeature([{ name: 'User', schema: UserSchema }])],
+//   imports: [MongooseModule.forFeature([{ name: 'User', schema: UserAuthSchema }])],
 //   controllers: [UserController],
 //   providers: [UserService],
 // })
